+++ static/js/starred_messages.js
@@ -1,10 +1,3 @@
-import $ from "jquery";
-
-import render_confirm_unstar_all_messages from "../templates/confirm_unstar_all_messages.hbs";
-import render_confirm_unstar_all_messages_in_topic from "../templates/confirm_unstar_all_messages_in_topic.hbs";
-
-import * as confirm_dialog from "./confirm_dialog";
-import * as message_flags from "./message_flags";
 import * as message_store from "./message_store";
 import * as stream_popover from "./stream_popover";
 import * as top_left_corner from "./top_left_corner";
@@ -82,36 +75,4 @@ export function rerender_ui() {
     stream_popover.hide_topic_popover();
     top_left_corner.update_starred_count(count);
     stream_popover.hide_starred_messages_popover();
-}
-
-export function confirm_unstar_all_messages() {
-    const modal_parent = $(".left-sidebar-modal-holder");
-    const html_body = render_confirm_unstar_all_messages();
-
-    confirm_dialog.launch({
-        parent: modal_parent,
-        html_heading: i18n.t("Unstar all messages"),
-        html_body,
-        html_yes_button: i18n.t("Unstar messages"),
-        on_click: message_flags.unstar_all_messages,
-    });
-}
-
-export function confirm_unstar_all_messages_in_topic(stream_id, topic) {
-    function on_click() {
-        message_flags.unstar_all_messages_in_topic(stream_id, topic);
-    }
-
-    const modal_parent = $(".left-sidebar-modal-holder");
-    const html_body = render_confirm_unstar_all_messages_in_topic({
-        topic,
-    });
-
-    confirm_dialog.launch({
-        parent: modal_parent,
-        html_heading: i18n.t("Unstar messages in topic"),
-        html_body,
-        html_yes_button: i18n.t("Unstar messages"),
-        on_click,
-    });
 }
