+++ tools/test-js-with-node
@@ -132,7 +132,7 @@ EXEMPT_FILES = {
     "static/js/settings_users.js",
     "static/js/setup.js",
     "static/js/spoilers.js",
-    "static/js/starred_messages.js",
+    "static/js/starred_messages_ui.js",
     "static/js/stream_bar.js",
     "static/js/stream_color.js",
     "static/js/stream_create.js",
